requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
PyMuPDF
Pillow
python-dotenv

sentence-transformers